Run Vitest suites as named projects
