+++ package.json
@@ -14,11 +14,11 @@
     "preview-app": "conc -r 'pnpm -F app run preview' 'pnpm -F nextjs run preview'",
     "preview-app-lite": "cross-env DISABLE_REFERENCE_LINKS=true pnpm run preview-app",
     "react18": "ariakit react18",
-    "test": "vitest",
+    "test": "vitest --project core --project node",
     "bench": "vitest bench --config vitest.bench.config.ts",
-    "test-react": "cross-env ARIAKIT_TEST_LOADER=react vitest",
+    "test-react": "vitest --project react",
     "test-react18": "ariakit react18 -- pnpm test-react",
-    "test-solid": "cross-env ARIAKIT_TEST_LOADER=solid vitest",
+    "test-solid": "vitest --project solid",
     "test-browser": "pnpm -F app test",
     "test-chrome": "pnpm -F app test-chrome",
     "test-chrome-headed": "pnpm -F app test-chrome-headed",
