+++ packages/ariakit-scripts/src/perf.test.ts
@@ -1,5 +1,3 @@
-// @vitest-environment node
-
 import { existsSync, mkdtempSync, readFileSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import path from "node:path";
