+++ vitest.react.setup.ts
@@ -1,0 +1,3 @@
+import { setupFrameworkTests } from "./vitest.framework.setup.ts";
+
+setupFrameworkTests("react");
